Add basic WeMo LED/Link support
This commit adds basic support for the WeMo LED bulbs that connect to the WeMo Link.
Because the bulbs must be accessed through the Link, the implementation is a little
different. Care has been taken to provide similar functionality to existing WeMo
Switches. This commit adds CLI and Python API compatibility.